\name{diaplt-package}
\alias{diaplt-package}
\alias{diaplt}
\docType{package}
\title{
\packageTitle{diaplt}
}
\description{
\packageDescription{diaplt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diaplt}
\packageIndices{diaplt}
This package contains beadsplot function.
}
\author{
\packageAuthor{diaplt}

Maintainer: \packageMaintainer{diaplt}
}
\references{
Beads Summary Plot of Ranges: \url{http://tomizonor.wordpress.com/2013/11/12/beads-plot/}
}
\keyword{ package }
\keyword{ hplot }
\seealso{
  \code{\link{beadsplot}}.
}
\examples{
beadsplot(Species~., iris)
}
